Reorganized invoicing DTOs
