--- LBH.AdultSocialCare.Api.CodeGenerator/Dockerfile
@@ -1,12 +1,0 @@
-FROM mcr.microsoft.com/dotnet/core/sdk:3.1
-
-WORKDIR /app
-
-# Copy csproj and restore as distinct layers
-COPY *.csproj ./
-RUN dotnet restore
-
-# Copy everything else and build
-COPY . ./
-RUN dotnet build -c Release -o out
-RUN dotnet publish -c Release -o out
