+++ LBH.AdultSocialCare.Api/V1/BusinessRules/Invoicing/Generators/CareChargeGenerator.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using HttpServices.Models.Requests;
 using LBH.AdultSocialCare.Api.Helpers;
+using LBH.AdultSocialCare.Api.V1.Domain.Common.Invoicing;
 
 namespace LBH.AdultSocialCare.Api.V1.BusinessRules.Invoicing.Generators
 {
