+++ LBH.AdultSocialCare.Api/V1/BusinessRules/Invoicing/Generators/FundedNursingCareGenerator.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using HttpServices.Models.Requests;
 using LBH.AdultSocialCare.Api.Helpers;
 using LBH.AdultSocialCare.Api.V1.AppConstants;
+using LBH.AdultSocialCare.Api.V1.Domain.Common.Invoicing;
 using LBH.AdultSocialCare.Api.V1.Domain.NursingCare;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.NursingCare;
