+++ LBH.AdultSocialCare.Api/V1/BusinessRules/Invoicing/Generators/IInvoiceItemsGenerator.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using HttpServices.Models.Requests;
+using LBH.AdultSocialCare.Api.V1.Domain.Common.Invoicing;
 
 namespace LBH.AdultSocialCare.Api.V1.BusinessRules.Invoicing.Generators
 {
