+++ LBH.AdultSocialCare.Api/V1/Domain/Common/Invoicing/AdditionalNeedsCost.cs
@@ -1,0 +1,11 @@
+using LBH.AdultSocialCare.Api.V1.Domain.NursingCare;
+
+namespace LBH.AdultSocialCare.Api.V1.Domain.Common.Invoicing
+{
+    public class AdditionalNeedsCost
+    {
+        public AdditionalNeedsPaymentTypeDomain AdditionalNeedsPaymentType { get; set; }
+
+        public decimal Cost { get; set; }
+    }
+}
