+++ LBH.AdultSocialCare.Api/V1/Domain/Common/Invoicing/BrokerageInfo.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace LBH.AdultSocialCare.Api.V1.Domain.Common.Invoicing
+{
+    public class BrokerageInfo
+    {
+        public decimal Core { get; set; }
+        public ICollection<AdditionalNeedsCost> AdditionalNeedsCosts { get; set; }
+    }
+}
