+++ LBH.AdultSocialCare.Api/V1/Factories/EntityToDomainFactory.cs
@@ -22,6 +22,7 @@ using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.ResidentialCarePackageR
 using System.Collections.Generic;
 using System.Linq;
 using LBH.AdultSocialCare.Api.V1.BusinessRules;
+using LBH.AdultSocialCare.Api.V1.Domain.Common.Invoicing;
 
 namespace LBH.AdultSocialCare.Api.V1.Factories
 {
